% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_form.R
\name{data_form}
\alias{data_form}
\title{Format user selected data}
\usage{
data_form(file_path)
}
\arguments{
\item{file_path}{parent path where the data exists with folder for each participant}
}
\value{
list object with each component giving compiled data for each data source
}
\description{
Format and join data for visualization
}
