% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev_vs_time.R
\name{fev1_vs_time}
\alias{fev1_vs_time}
\title{Trelliscope of FEV1 over time}
\usage{
fev1_vs_time(spiro_data)
}
\arguments{
\item{spiro_data}{data.frame with at minimum columns giving "fev1", "when_time", "Sample_Info", and "when_date"}
}
\value{
plotly display showing FEV1 over time. If more than one subject (participant and season) is in the data division then subject data is connected and colored differently.
}
\description{
Panel function for FEV1 over time
}
