% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fev_vs_time.R
\name{fev1_vs_time_cog}
\alias{fev1_vs_time_cog}
\title{Trelliscope cognostics of FEV1 over time}
\usage{
fev1_vs_time_cog(spiro_data)
}
\arguments{
\item{spiro_data}{data.frame with at minimum columns giving "fev1", "when_time", "Sample_Info", and "when_date"}
}
\value{
tibble containing cognostics for trelliscopejs
}
\description{
Cognostic function for FEV1 over time
}
