% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pah_conversion_table.R
\name{pah_conversion_table}
\alias{pah_conversion_table}
\title{PAH Conversion Table}
\format{A 63x4 data.frame giving FSES parameter id, chemical name, CAS code, and molecular weight}
\description{
Gives molecular weight (grams/mol) of each Chemical
}
