% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whenTime_format.R
\name{whenTime_format}
\alias{whenTime_format}
\title{Format date correctly from the 'when_time' column}
\usage{
whenTime_format(time_vector)
}
\arguments{
\item{time_vector}{current format is DayofWeek_Month_Day_Year_HH:MM:SS_TimeZone}
}
\description{
Format date to match POSIX accepted formats
}
